requirements.txt
tensorflow==2.15.0
rembg==2.0.57
numpy==1.26.4
matplotlib==3.9.0
fastapi==0.111.0
pillow==10.3.0
uvicorn==0.30.1
ultralytics==8.2.50
ultralyticsplus
